Replace url user security checks by controller users checks from secu…

+++ src/main/java/org/barino3d/controllers/ApplicationController.java
@@ -17,6 +17,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 import java.util.List;
 
+import static org.barino3d.controllers.ControllersHelper.checkUserIsOwnerOfTheApplication;
+import static org.barino3d.controllers.ControllersHelper.getCurrentUserEmail;
+
 @Controller
 @AllArgsConstructor(onConstructor = @____(@Autowired))
 public class ApplicationController {
@@ -25,22 +28,25 @@ public class ApplicationController {
     private final CommandService commandService;
     private final UserService userService;
 
-    @PostMapping("user/{userId}/application")
-    public String addApplication(@PathVariable String userId, @ModelAttribute(value = "newApp") Application application) {
+    @PostMapping("application")
+    public String addApplication(@ModelAttribute(value = "newApp") Application application) {
+        String userId = userService.findByEmail(getCurrentUserEmail()).getId();
         application.setUser(userService.findById(userId));
         applicationService.save(application);
-        return "redirect:/user/" + userId + "/application/" + application.getId();
+        return "redirect:/" + "application/" + application.getId();
     }
 
-    @GetMapping("user/{userId}/application/{id}")
-    public String getApplication(@PathVariable String userId, @PathVariable String id, Model model) {
-        UserEntity user = userService.findById(userId);
+    @GetMapping("application/{id}")
+    public String getApplication(@PathVariable String id, Model model) {
+        String userId = userService.findByEmail(getCurrentUserEmail()).getId();
+        UserEntity currentUser = userService.findById(userId);
+        checkUserIsOwnerOfTheApplication(currentUser, applicationService.findById(id));
         UserEntity libraryUser = userService.findById("600eb29748058538a3274fee");
-        List<Application> applications = applicationService.findAllByUser(user);
+        List<Application> applications = applicationService.findAllByUser(currentUser);
         List<Application> libraryApplications = applicationService.findAllByUser(libraryUser);
         final Application currentApplication = applicationService.findById(id);
         List<Command> commands = currentApplication.getCommands();
-        model.addAttribute("user", user);
+        model.addAttribute("currentUser", currentUser);
         model.addAttribute("currentApplication", currentApplication);
         model.addAttribute("applications", applications);
         model.addAttribute("libraryApplications", libraryApplications);
@@ -51,19 +57,22 @@ public class ApplicationController {
         return "index";
     }
 
-    @PostMapping("user/{userId}/application/{id}/delete")
-    public String deleteApplication(@PathVariable String userId, @PathVariable String id) {
+    @PostMapping("application/{id}/delete")
+    public String deleteApplication(@PathVariable String id) {
+        String userId = userService.findByEmail(getCurrentUserEmail()).getId();
+        UserEntity currentUser = userService.findById(userId);
+        final Application application = applicationService.findById(id);
+        checkUserIsOwnerOfTheApplication(currentUser, application);
         if (applicationService.findAllByUser(userService.findById(userId)).size() == 1) {
             return "redirect:/";
         }
-        final Application application = applicationService.findById(id);
         commandService.deleteAllByApplication(application);
         applicationService.delete(application);
-        return "redirect:/user/" + userId + "/application/" + applicationService.findAllByUser(userService.findById(userId)).get(0).getId();
+        return "redirect:/" + "application/" + applicationService.findAllByUser(userService.findById(userId)).get(0).getId();
     }
 
-    @PostMapping("user/{userId}/application/{id}/library")
-    public String getCommandsFromLibraryApplication(@PathVariable String userId, @PathVariable String id, @ModelAttribute(value = "libraryApplication") Application libraryApplication) {
+    @PostMapping("application/{id}/library")
+    public String getCommandsFromLibraryApplication(@PathVariable String id, @ModelAttribute(value = "libraryApplication") Application libraryApplication) {
         final Application userApplication = applicationService.findById(id);
         final Application libraryApplicationFromRepo = applicationService.findById(libraryApplication.getId());
         libraryApplicationFromRepo.getCommands().forEach(
@@ -72,6 +81,7 @@ public class ApplicationController {
                 }
         );
         applicationService.save(userApplication);
-        return "redirect:/user/" + userId + "/application/" + id;
+        return "redirect:/" + "application/" + id;
     }
+
 }
